backend/requirements.txt
tornado==6.4.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
redis==5.0.1
lz4==4.3.3
requests==2.31.0
numpy==1.26.4
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
websocket-client==1.6.4
